requirements.txt
pandas
numpy
psycopg2-binary
pgvector
sqlalchemy
anthropic
sentence-transformers
scikit-learn
hdbscan
umap-learn
streamlit
plotly
python-dotenv
requests
beautifulsoup4
scipy
pdfplumber
openpyxl
lxml
tqdm
